update file structure for a component
