+++ src/forms/Input/Input.tsx
@@ -1,27 +1,8 @@
 import * as React from "react";
-import styled, { css } from "styled-components";
 
 import { Spacer } from "../../Spacer";
-import { TextCss } from "../TextCss";
-
-export type InputProps = {
-  hasError?: boolean;
-  isDisabled?: boolean;
-  onChange: (value: string) => void;
-  /** React Node that will be placed before the text, but inside of the input border. */
-  prefixNode?: React.ReactNode;
-  /** React Node that will be placed after the text, but inside of the input border. */
-  suffixNode?: React.ReactNode;
-  /**
-   * Specifies visual details of the input component. If none provided,
-   * then `InputThemeDefault` will be used.
-   */
-  inputElementProps?: React.HTMLAttributes<HTMLInputElement>;
-  value: string;
-  testId?: string;
-  placeholder?: string;
-  type?: string;
-};
+import { InputContainer } from "./styles";
+import { InputProps } from "./types";
 
 export const Input: React.FC<InputProps> = ({
   hasError,
@@ -62,67 +43,6 @@ export const Input: React.FC<InputProps> = ({
     </InputContainer>
   );
 };
-
-const InputContainer = styled.div`
-  ${(props) => {
-    const {
-      form: {
-        input: {
-          backgroundColor,
-          disabledBackgroundColor,
-          errorColor,
-          height,
-          borderColor,
-          focusBorderColor,
-          placeholderColor,
-        },
-      },
-    } = props.theme;
-
-    return css`
-      display: flex;
-      align-items: center;
-      background: ${backgroundColor};
-      border: 1px solid ${borderColor};
-      cursor: text;
-      height: ${height};
-      padding: 0;
-
-      &.is-disabled {
-        background: ${disabledBackgroundColor};
-      }
-
-      :focus,
-      :focus-within {
-        border: 1px solid ${focusBorderColor};
-      }
-
-      &.has-error {
-        border: 1px solid ${errorColor};
-      }
-
-      input {
-        padding: 0;
-        background: none;
-        cursor: inherit;
-        border: 0;
-        appearance: none;
-        outline: none;
-        flex: 1;
-        ${TextCss}
-
-        :invalid {
-          appearance: none;
-          box-shadow: none;
-        }
-
-        ::placeholder {
-          color: ${placeholderColor};
-        }
-      }
-    `;
-  }}
-`;
 
 const computeInputClasses = (
   hasError: boolean = false,
