+++ src/forms/Input/index.ts
@@ -1,0 +1,4 @@
+export * from "./Input";
+export * from "./themes";
+export * from "./types";
+export * from "./styles";
